backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<3.16
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
mysqlclient>=2.2
python-decouple>=3.8
django-filter>=23.5,<25.0
